rename side panel to tool panel
