+++ DESIGN.md
@@ -122,11 +122,11 @@ The main application layout in `index.html` is based on [CSS grid](https://devel
 
 - `<div id="app-container">` serves as the main grid container and defines template areas.
 - |--- `<main class="content">` serves as the main tab container with Bible text, search results or verse list content in each tab.
-- |--- `<aside id="side-panel">` is a container for quick reference info such as tag list and Strong's dictionary.
-- |--- `<footer id="bottom-panel">` is a container for detailed information regarding selected verse (interlinear view) or clicked Cross reference link.
+- |--- `<aside id="tool-panel">` is a container for quick reference info such as tag list and Strong's dictionary.
+- |--- `<footer id="detail-panel">` is a container for detailed information regarding selected verse (interlinear view) or clicked Cross reference link.
 
 Please note, that each tab in `<main class="content">` has a subgrid for layout in that particular tab.
-Dictionary view on smaller screen sizes (tablets) will be displayed in the `bottom-panel` area.
+Tool panel in portrait mode will be displayed on the bottom of the screen.
 
 
 <a name='db-schema'></a>
