+++ app/frontend/components/tool_panel/panel_buttons.js
