+++ index.html
@@ -117,7 +117,7 @@
           <i class="fa fa-book"></i>
         </button>
       </panel-buttons>
-      <aside id="side-panel" class="with-tags">
+      <aside id="tool-panel" class="with-tags">
         <section id="bible-browser-toolbox"></section>
         <section id="dictionary-info-box" class="accordion-panel">
           <h3>
